Integrate new components.
