--- src/components/Office/ReviewServiceItems/NewReviewServiceItems.module.scss
@@ -1,92 +1,0 @@
-@import '../../../shared/styles/basics';
-@import '../../../shared/styles/mixins';
-@import '../../../shared/styles/colors';
-
-.ReviewServiceItems {
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-
-  box-shadow: -4px 0px 8px 0px rgba(0, 0, 0, 0.1);
-
-  .top button[type='button'] {
-    float: right;
-    // on mobile the button becomes flexed so must override centering
-    justify-content: flex-end;
-    @include u-padding(0);
-    svg {
-      @include u-width(2);
-      @include u-height(2);
-      path {
-        fill: $primary;
-      }
-    }
-  }
-
-  .eyebrowTitle {
-    @include u-text('base');
-    @include u-font-size('body', '3xs');
-    @include u-padding-top(1);
-  }
-
-  .header {
-    @include u-margin-top(1);
-    @include u-margin-bottom(0);
-  }
-
-  .top {
-    @include u-padding(3);
-    @include u-padding-left(4);
-    @include u-bg('white');
-    @include u-border-bottom('2px');
-    @include u-border-bottom('gray-10');
-    @include u-height('auto');
-  }
-
-  .body {
-    @include u-padding-y(3);
-    @include u-padding-x(4);
-    background-color: $bg-gray;
-    @include u-overflow-y('auto');
-    flex: 1 1 0;
-    flex-basis: 0;
-
-    :global .usa-alert {
-      @include u-margin-bottom(2);
-    }
-  }
-
-  .bottom {
-    @include u-padding-x(3);
-    @include u-padding-y(1);
-    @include u-bg('white');
-
-    @include u-border-top('2px');
-    @include u-border-top('gray-10');
-    @include u-position('relative');
-
-    display: flex;
-    align-items: center;
-
-    .totalApproved {
-      display: flex;
-      flex-direction: column;
-      flex-grow: 1;
-      text-align: right;
-
-      .totalLabel {
-        @include u-font-size('body', '3xs');
-        @include u-text('bold');
-        @include u-padding-top(1);
-        color: $base-darker;
-      }
-
-      .totalAmount {
-        @include u-font-size('body', 'xs');
-        @include u-padding-bottom(1);
-        color: $base-darkest;
-        flex-grow: 1;
-      }
-    }
-  }
-}
