+++ src/components/Office/ReviewServiceItems/ServiceItemCard.stories.jsx
@@ -1,4 +1,6 @@
 import React from 'react';
+import { expect } from '@storybook/jest';
+import { within, userEvent } from '@storybook/testing-library';
 
 import testParams from '../ServiceItemCalculations/serviceItemTestParams';
 
@@ -26,6 +28,26 @@ export const Basic = (args) => (
     patchPaymentServiceItem={args.patchPaymentServiceItem}
   />
 );
+
+export const EmptyRejectionReasonError = (args) => (
+  <ServiceItemCard
+    mtoServiceItemName={serviceItemCodes.CS}
+    amount={999.99}
+    patchPaymentServiceItem={args.patchPaymentServiceItem}
+    status={PAYMENT_SERVICE_ITEM_STATUS.DENIED}
+  />
+);
+
+EmptyRejectionReasonError.play = async ({ canvasElement }) => {
+  const canvas = within(canvasElement);
+
+  await expect(canvas.getByRole('textbox', { name: 'Reason for rejection' })).toBeInTheDocument();
+  await expect(canvas.getByText('Reject')).toBeInTheDocument();
+
+  // type, then clear, then blur
+  await userEvent.type(canvas.getByRole('textbox', { name: 'Reason for rejection' }), 'a{backspace}');
+  await userEvent.click(canvas.getByText('Reject'), undefined, { skipPointerEventsCheck: true });
+};
 
 export const HHG = (args) => (
   <ServiceItemCard
