+++ src/components/Office/ReviewServiceItems/ServiceItemCard.test.jsx
@@ -1,6 +1,7 @@
 /* eslint-disable react/jsx-props-no-spreading */
 import React from 'react';
-import { act, fireEvent, render, screen, waitFor } from '@testing-library/react';
+import { act, render, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 import { mount } from 'enzyme';
 
 import testParams from '../ServiceItemCalculations/serviceItemTestParams';
@@ -20,10 +21,12 @@ const basicServiceItemCard = {
   status: PAYMENT_SERVICE_ITEM_STATUS.REQUESTED,
   patchPaymentServiceItem: jest.fn(),
 };
+
 const reviewedBasicServiceItemCard = {
   ...basicServiceItemCard,
   requestComplete: true,
 };
+
 const needsReviewServiceItemCard = {
   id: '1',
   mtoShipmentType: SHIPMENT_OPTIONS.HHG,
@@ -39,6 +42,7 @@ const needsReviewServiceItemCard = {
   paymentServiceItemParams: testParams.FuelSurchage,
   patchPaymentServiceItem: jest.fn(),
 };
+
 const additionalDaySITServiceItemCard = {
   ...needsReviewServiceItemCard,
   mtoServiceItemName: serviceItemCodes.DOASIT,
@@ -54,107 +58,49 @@ const additionalDaySITServiceItemCard = {
     totalSITEndDate: '2021-09-08',
   },
 };
+
 const reviewedServiceItemCard = {
   ...needsReviewServiceItemCard,
   requestComplete: true,
 };
+
 const canceledShipmentServiceItemCard = {
   ...needsReviewServiceItemCard,
   mtoShipmentModificationType: shipmentModificationTypes.CANCELED,
 };
+
 const divertedShipmentServiceItemCard = {
   ...needsReviewServiceItemCard,
   mtoShipmentModificationType: shipmentModificationTypes.DIVERSION,
 };
+
 describe('ServiceItemCard component', () => {
   describe('when payment request needs reviewed', () => {
     const wrapper = mount(<ServiceItemCard {...needsReviewServiceItemCard} />);
     it('toggles pricer calculations when button is clicked', () => {
       const toggleButton = wrapper.find('button[data-testid="toggleCalculations"]');
       expect(toggleButton.text()).toEqual('Show calculations');
+
       act(() => {
         toggleButton.simulate('click');
       });
       wrapper.update();
+
       expect(toggleButton.text()).toEqual('Hide calculations');
       expect(wrapper.find('ServiceItemCalculations').exists()).toBe(true);
+
       act(() => {
         toggleButton.simulate('click');
       });
       wrapper.update();
+
       expect(toggleButton.text()).toEqual('Show calculations');
       expect(wrapper.find('ServiceItemCalculations').exists()).toBe(false);
     });
+
     it('does not render calculations toggle when the service item calculations are not implemented', () => {
       const component = mount(<ServiceItemCard {...basicServiceItemCard} />);
       expect(component.find('button[data-testid="toggleCalculations"]').exists()).toBe(false);
-    });
-
-    // using react testing library to test dom interactions
-    it('disables the save button when rejection reason is empty', async () => {
-      render(<ServiceItemCard {...basicServiceItemCard} />);
-
-      fireEvent.click(screen.getByTestId('rejectRadio'));
-
-      await waitFor(() => {
-        expect(screen.getByTestId('rejectionSaveButton')).toHaveAttribute('disabled');
-      });
-    });
-
-    // using react testing library to test dom interactions
-    it('shows edit reason link after saving', async () => {
-      render(<ServiceItemCard {...basicServiceItemCard} />);
-
-      // Click on reject radio and fill in text area
-      fireEvent.click(screen.getByTestId('rejectRadio'));
-      fireEvent.change(screen.getByTestId('textarea'), {
-        target: { value: 'Rejected just because.' },
-      });
-
-      await waitFor(() => {
-        expect(screen.getByTestId('rejectionSaveButton').hasAttribute('disabled')).toBeFalsy();
-      });
-
-      // Save
-      fireEvent.click(screen.getByTestId('rejectionSaveButton'));
-
-      await waitFor(() => {
-        expect(screen.getByTestId('editReasonButton')).toBeTruthy();
-      });
-      await waitFor(() => {
-        expect(screen.getByTestId('rejectionReasonReadOnly').textContent).toBe('Rejected just because.');
-      });
-    });
-
-    // using react testing library to test dom interactions
-    it('edits the rejection reason', async () => {
-      const data = {
-        ...basicServiceItemCard,
-        status: PAYMENT_SERVICE_ITEM_STATUS.DENIED,
-        rejectionReason: 'Rejected just because.',
-      };
-      render(<ServiceItemCard {...data} />);
-
-      await waitFor(() => {
-        expect(screen.getByTestId('editReasonButton')).toBeTruthy();
-      });
-      await waitFor(() => {
-        expect(screen.getByTestId('rejectionReasonReadOnly').textContent).toBe('Rejected just because.');
-      });
-
-      // Click on Edit reason button, edit text area and save
-      fireEvent.click(screen.getByTestId('editReasonButton'));
-      fireEvent.change(screen.getByTestId('textarea'), {
-        target: { value: 'Edited rejection reason.' },
-      });
-      fireEvent.click(screen.getByTestId('rejectionSaveButton'));
-
-      await waitFor(() => {
-        expect(screen.getByTestId('editReasonButton')).toBeTruthy();
-      });
-      await waitFor(() => {
-        expect(screen.getByTestId('rejectionReasonReadOnly').textContent).toBe('Edited rejection reason.');
-      });
     });
 
     it('displays the Days In SIT information for additional day service items', () => {
@@ -164,41 +110,96 @@ describe('ServiceItemCard component', () => {
     });
   });
 
+  describe('when Reject is selected', () => {
+    it('the component displays correctly', async () => {
+      render(<ServiceItemCard {...basicServiceItemCard} />);
+
+      await waitFor(() => {
+        expect(screen.getByRole('heading', { level: 3, name: 'BASIC SERVICE ITEMS' })).toBeInTheDocument();
+      });
+      const approveButton = screen.getByLabelText('Approve');
+      const rejectButton = screen.getByLabelText('Reject');
+      expect(screen.getByLabelText('Reject')).toBeInTheDocument();
+      expect(approveButton).toBeInTheDocument();
+
+      await userEvent.click(rejectButton);
+      expect(rejectButton).toBeChecked();
+      expect(screen.queryByText('Add a reason why this service item is rejected')).not.toBeInTheDocument();
+      expect(screen.getByLabelText('Reason for rejection')).toBeInTheDocument();
+    });
+    describe('when a reason is added', () => {
+      it('Approve is selected, and Reject is reselected, the reason is cleared, and no error appears', async () => {
+        render(<ServiceItemCard {...basicServiceItemCard} />);
+
+        const approveButton = screen.getByLabelText('Approve');
+        const rejectButton = screen.getByLabelText('Reject');
+
+        await userEvent.click(rejectButton);
+        expect(screen.getByLabelText(/Reason for rejection/)).toBeInTheDocument();
+        const reason = 'why it was rejected';
+        await userEvent.type(screen.getByLabelText(/Reason for rejection/), reason);
+        await userEvent.click(approveButton);
+        await userEvent.click(rejectButton);
+        expect(screen.queryByText('Add a reason why this service item is rejected')).not.toBeInTheDocument();
+        expect(screen.queryByText(reason)).not.toBeInTheDocument();
+      });
+      it('and removed, and the textbox is blurred, an error is shown', async () => {
+        // TODO: the same error will show if the form is submitted without a reason; test for this in integration
+        render(<ServiceItemCard {...basicServiceItemCard} />);
+        const rejectButton = screen.getByLabelText('Reject');
+        await userEvent.click(rejectButton);
+        expect(screen.getByLabelText(/Reason for rejection/)).toBeInTheDocument();
+        await userEvent.type(screen.getByLabelText(/Reason for rejection/), 'a{backspace}');
+        await userEvent.click(rejectButton);
+        expect(screen.queryByText('Add a reason why this service item is rejected')).toBeInTheDocument();
+      });
+    });
+  });
+
   describe('when payment request has been reviewed', () => {
     const wrapper = mount(<ServiceItemCard {...reviewedServiceItemCard} />);
     it('toggles pricer calculations when button is clicked', () => {
       const toggleButton = wrapper.find('button[data-testid="toggleCalculations"]');
       expect(toggleButton.text()).toEqual('Show calculations');
+
       act(() => {
         toggleButton.simulate('click');
       });
       wrapper.update();
+
       expect(toggleButton.text()).toEqual('Hide calculations');
       expect(wrapper.find('ServiceItemCalculations').exists()).toBe(true);
+
       act(() => {
         toggleButton.simulate('click');
       });
       wrapper.update();
+
       expect(toggleButton.text()).toEqual('Show calculations');
       expect(wrapper.find('ServiceItemCalculations').exists()).toBe(false);
     });
+
     it('does not render calculations toggle when the service item calculations are not implemented', () => {
       const component = mount(<ServiceItemCard {...reviewedBasicServiceItemCard} />);
       expect(component.find('button[data-testid="toggleCalculations"]').exists()).toBe(false);
     });
+
     it('does not display days in SIT info for additional day service items', () => {
       const reviewedDOASIT = { ...additionalDaySITServiceItemCard, requestComplete: true };
+
       render(<ServiceItemCard {...reviewedDOASIT} />);
       expect(screen.queryByText('SIT days invoiced')).not.toBeInTheDocument();
       expect(screen.queryByText('DaysInSITAllowance')).not.toBeInTheDocument();
     });
   });
+
   describe('When a service item has a shipment that was canceled ', () => {
     const component = mount(<ServiceItemCard {...canceledShipmentServiceItemCard} />);
     it('there is a canceled tag displayed', () => {
       expect(component.find('ShipmentModificationTag').text()).toBe(shipmentModificationTypes.CANCELED);
     });
   });
+
   describe('When a service item has a shipment that was diverted ', () => {
     const component = mount(<ServiceItemCard {...divertedShipmentServiceItemCard} />);
     it('there is a diversion tag displayed', () => {
